Prepend modes to the channel topic line
This is based off the patch attached to a feature request:
"[PATCH] Channel mode in titlebar"
https://bugs.quassel-irc.org/issues/1161

Channel modes are prepended to the topic and removed when editing
the topic. This is a hacky way of showing the modes, but it does
the trick for now. In the future I'll look at a proper widget for
modes display, beside the topic widget.